moving items and dialogs to subfolders to keep the project well organized
