+++ Lib/defconQt/dialogs/addAnchorDialog.py
