+++ Lib/defconQt/dialogs/addLayerDialog.py
