+++ Lib/defconQt/dialogs/gotoDialog.py
