+++ Lib/defconQt/dialogs/layerActionsDialog.py
