+++ Lib/defconQt/items/componentItem.py
