+++ Lib/defconQt/items/handleLineItem.py
