+++ Lib/defconQt/items/offCurvePointItem.py
