+++ Lib/defconQt/items/onCurvePointItem.py
