+++ Lib/defconQt/items/pixmapItem.py
