+++ Lib/defconQt/items/startPointItem.py
