+++ Lib/defconQt/items/vGuidelinesTextItem.py
