+++ Lib/defconQt/layerSetList.py
@@ -4,7 +4,7 @@ from PyQt5.QtWidgets import (
     QApplication, QListWidget, QListWidgetItem, QAbstractItemView, QAction,
     QColorDialog)
 from defconQt import icons_db  # noqa
-from defconQt.addLayerDialog import AddLayerDialog
+from defconQt.dialogs.addLayerDialog import AddLayerDialog
 
 class LayerSetList(QListWidget):
